DevW25Nxx update

+++ Device/Dev_W25Nxx.h
@@ -171,7 +171,7 @@ typedef struct
 typedef struct
 {
     DevW25Nxx_ProdType_List prod_type;
-    uint16_t prod_code;
+    uint32_t prod_code;
     uint8_t *tmp_buf;
 
     uint16_t (*bus_tx)(uint8_t *p_data, uint16_t len, uint32_t time_out);
